Rename auth controller to camelCase
